backend/historian_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.16.0
google-adk
google-genai
